+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/__init__.py
@@ -3,12 +3,25 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+from azure.cli.core import AzCommandsLoader
+
 import azure.cli.command_modules.acr._help  # pylint: disable=unused-import
 
 
-def load_params(_):
-    import azure.cli.command_modules.acr._params  # pylint: disable=redefined-outer-name, unused-variable
+class ACRCommandsLoader(AzCommandsLoader):
+
+    def __init__(self, cli_ctx=None):
+        super(ACRCommandsLoader, self).__init__(cli_ctx=cli_ctx,
+                                                min_profile='2017-03-10-profile')
+
+    def load_command_table(self, args):
+        from azure.cli.command_modules.acr.commands import load_command_table
+        load_command_table(self, args)
+        return self.command_table
+
+    def load_arguments(self, command):
+        from azure.cli.command_modules.acr._params import load_arguments
+        load_arguments(self, command)
 
 
-def load_commands():
-    import azure.cli.command_modules.acr.commands  # pylint: disable=redefined-outer-name, unused-variable
+COMMAND_LOADER_CLS = ACRCommandsLoader