+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.0.21
+++++++
+* add korea south and korea central to preview regions
+* add `az aks install-connector` and `az aks remove-connector` commands
+
 2.0.20
 ++++++
 * `acs create`: emit out an actionable error if provisioning application failed for lack of permissions